Move `HDSeedFingerprint` to `zcash_keys`
