+++ zcash_keys/Cargo.toml
@@ -42,6 +42,7 @@ memuse.workspace = true
 tracing.workspace = true
 
 # - Secret management
+secrecy.workspace = true
 subtle.workspace = true
 
 # - Shielded protocols
@@ -60,6 +61,9 @@ document-features.workspace = true
 
 # - Encodings
 byteorder = { workspace = true, optional = true }
+
+# - Digests
+blake2b_simd = { workspace = true }
 
 [dev-dependencies]
 hex.workspace = true
